+++ Plugins/Bugsnag/Source/Bugsnag/Private/Apple/ApplePlatformConfiguration.cpp
@@ -12,9 +12,13 @@
 #import <Bugsnag/BugsnagEndpointConfiguration.h>
 #import <Bugsnag/BugsnagErrorTypes.h>
 
-#import <BugsnagPrivate/BSGJSONSerialization.h>
 #import <BugsnagPrivate/BugsnagConfiguration+Private.h>
 #import <BugsnagPrivate/BugsnagNotifier.h>
+
+static BSGTelemetryOptions GetTelemetryTypes(EBugsnagTelemetryTypes Value)
+{
+	return (EnumHasAllFlags(Value, EBugsnagTelemetryTypes::InternalErrors) ? BSGTelemetryInternalErrors : 0);
+}
 
 static BSGThreadSendPolicy GetThreadSendPolicy(EBugsnagSendThreadsPolicy Policy)
 {
@@ -110,6 +114,8 @@ BugsnagConfiguration* FApplePlatformConfiguration::Configuration(const TSharedRe
 	CocoaConfig.maxPersistedSessions = Configuration->GetMaxPersistedSessions();
 
 	CocoaConfig.persistUser = Configuration->GetPersistUser();
+
+	CocoaConfig.telemetry = GetTelemetryTypes(Configuration->GetTelemetry());
 
 	if (Configuration->GetReleaseStage().IsSet())
 	{